hands-on-2021-main/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=eigen
absl-py=0.12.0=py38haa95532_0
aiohttp=3.7.4=py38h2bbff1b_1
astunparse=1.6.3=py_0
async-timeout=3.0.1=py38haa95532_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py38haa95532_0
brotli-python=1.0.9=py38h82bb817_2
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2021.1.19=haa95532_1
cachetools=4.2.1=pyhd3eb1b0_0
certifi=2020.12.5=py38haa95532_0
cffi=1.14.5=py38hcd4344a_0
chardet=3.0.4=py38haa95532_1003
click=7.1.2=pyhd3eb1b0_0
colorama=0.4.4=pyhd3eb1b0_0
coverage=5.5=py38h2bbff1b_2
cryptography=3.4.6=py38h71e12ea_0
cycler=0.10.0=py38_0
cython=0.29.22=py38hd77b12b_0
dash=1.19.0=pyhd3eb1b0_0
dash-bootstrap-components=0.11.3=pyh44b312d_0
dash-core-components=1.3.1=py_0
dash-html-components=1.0.1=py_0
dash-renderer=1.1.2=py_0
dash-table=4.4.1=py_0
decorator=4.4.2=pyhd3eb1b0_0
flask=1.1.2=pyhd3eb1b0_0
flask-compress=1.9.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
future=0.18.2=py38_1
gast=0.4.0=py_0
google-auth=1.28.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.3=pyhd3eb1b0_0
google-pasta=0.2.0=py_0
grpcio=1.36.1=py38hc60d5dd_1
h5py=2.10.0=py38h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.7.3=py38haa95532_1
intel-openmp=2020.2=254
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.21.0=py38hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
itsdangerous=1.1.0=pyhd3eb1b0_0
jedi=0.17.0=py38_0
jinja2=2.11.3=pyhd3eb1b0_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=hb83a4c4_2
jupyter_client=6.1.7=py_0
jupyter_core=4.7.1=py38haa95532_0
keras=2.4.3=0
keras-applications=1.0.8=py_1
keras-base=2.4.3=py_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.1=py38hd77b12b_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.14.0=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.2.0=he0120a3_0
lz4-c=1.9.3=h2bbff1b_0
markdown=3.3.4=py38haa95532_0
markupsafe=1.1.1=py38he774522_0
matplotlib=3.3.4=py38haa95532_0
matplotlib-base=3.3.4=py38h49ac443_0
mkl=2020.2=256
mkl-service=2.3.0=py38h196d8e1_0
mkl_fft=1.3.0=py38h46781fe_0
mkl_random=1.1.1=py38h47e9c7a_0
multidict=5.1.0=py38h2bbff1b_2
numpy=1.19.2=py38hadc3359_0
numpy-base=1.19.2=py38ha3acd2a_0
oauthlib=3.1.0=py_0
olefile=0.46=py_0
openssl=1.1.1k=h2bbff1b_0
opt_einsum=3.1.0=py_0
pandas=1.2.3=py38hf11a4ad_0
parso=0.8.1=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.2=py38h4fa10fc_0
pip=21.0.1=py38haa95532_0
plotly=4.14.3=pyhd3eb1b0_0
prompt-toolkit=3.0.8=py_0
protobuf=3.14.0=py38hd77b12b_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pygments=2.8.1=pyhd3eb1b0_0
pyjwt=1.7.1=py38_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38ha925a31_4
pyreadline=2.1=py38_1
pysocks=1.7.1=py38haa95532_0
python=3.8.8=hdbf39b2_4
python-dateutil=2.8.1=pyhd3eb1b0_0
python_abi=3.8=1_cp38
pytz=2021.1=pyhd3eb1b0_0
pywin32=227=py38he774522_1
pyyaml=5.4.1=py38h2bbff1b_1
pyzmq=20.0.0=py38hd77b12b_1
qt=5.9.7=vc14h73c81de_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
retrying=1.3.3=py_2
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=0.24.1=py38hf11a4ad_0
scipy=1.6.1=py38h14eb087_0
seaborn=0.11.1=pyhd3eb1b0_0
setuptools=52.0.0=py38haa95532_0
sip=4.19.13=py38ha925a31_0
six=1.15.0=py38haa95532_0
sqlite=3.35.2=h2bbff1b_0
tensorboard=2.4.0=pyhc547734_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.3.0=mkl_py38h8c0d9a2_0
tensorflow-base=2.3.0=eigen_py38h75a453f_0
tensorflow-estimator=2.3.0=pyheb71bc4_0
termcolor=1.1.0=py38haa95532_1
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_0
tornado=6.1=py38h2bbff1b_0
traitlets=5.0.5=pyhd3eb1b0_0
typing-extensions=3.7.4.3=hd3eb1b0_0
typing_extensions=3.7.4.3=pyh06a4308_0
urllib3=1.26.4=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=py_0
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38_0
wrapt=1.12.1=py38he774522_1
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
yarl=1.6.3=py38h2bbff1b_0
zeromq=4.3.3=ha925a31_3
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
